Creating JVM Reader File
To avoid overload the controller file, this
change creates a file for reading purposes and
a file to controlling.
